refactor: Split iterator.rs into module hierarchy (Issue #1144)
Refactor the 1,561-line iterator.rs file into a well-organized module structure:

**Module Structure:**
- iterator/mod.rs: Core trait, documentation, and re-exports
- iterator/scan.rs: TableScanIterator implementation
- iterator/filter.rs: FilterIterator implementation
- iterator/projection.rs: ProjectionIterator implementation
- iterator/join.rs: LazyNestedLoopJoin implementation
- iterator/tests/: Test suite organized by functionality
  - tests/basic.rs: Scan, filter, and projection tests
  - tests/join.rs: Join operation tests
  - tests/phase_c.rs: End-to-end integration tests

**Benefits:**
- Each file now under 300 lines (most under 150)
- Clear separation of concerns by iterator type
- Tests separated from implementation (~750 lines)
- Improved navigability and maintainability
- No public API changes - all items re-exported from mod.rs

**Verification:**
- All 19 iterator tests passing
- No regression in existing functionality
- Documentation and Phase C integration notes preserved
