--- crates/executor/src/select/iterator.rs
@@ -1,1561 +1,0 @@
-//! Lazy iterator-based query execution infrastructure
-//!
-//! This module provides a foundation for streaming query execution using iterators
-//! instead of materializing results. This reduces memory usage and enables early
-//! termination for LIMIT queries.
-//!
-//! # Architecture
-//!
-//! The core trait is `RowIterator`, which extends `Iterator<Item = Result<Row, ExecutorError>>`
-//! with additional query-specific methods. All query operators (scan, filter, project, join)
-//! are implemented as iterators that can be composed.
-//!
-//! # Benefits
-//!
-//! - **Memory efficiency**: O(max_single_table) instead of O(product)
-//! - **Streaming**: Rows flow through pipeline without buffering
-//! - **Early termination**: LIMIT 10 only computes 10 rows
-//! - **Composability**: Iterators naturally chain together
-//!
-//! # Example
-//!
-//! ```rust,ignore
-//! // Create a table scan iterator
-//! let scan_iter = TableScanIterator::new(schema, rows);
-//!
-//! // Add a filter
-//! let filter_iter = FilterIterator::new(scan_iter, predicate, evaluator);
-//!
-//! // Add a projection
-//! let project_iter = ProjectionIterator::new(filter_iter, projection_fn);
-//!
-//! // Consume only what we need (e.g., LIMIT 10)
-//! for row in project_iter.take(10) {
-//!     println!("{:?}", row?);
-//! }
-//! ```
-//!
-//! # Phase C Integration (Proof of Concept)
-//!
-//! The `build_simple_query_iterator()` function demonstrates how to build an iterator
-//! pipeline for simple SELECT queries (without ORDER BY, DISTINCT, or window functions).
-//! This serves as a proof-of-concept for full integration into the executor.
-
-use crate::{errors::ExecutorError, evaluator::CombinedExpressionEvaluator, schema::CombinedSchema};
-
-/// Core trait for row-producing iterators in the query execution pipeline
-///
-/// This trait extends the standard Iterator trait with query-specific metadata
-/// and methods. All query operators (scans, filters, joins, projections) implement
-/// this trait to enable composable, streaming query execution.
-///
-/// # Why not just use Iterator?
-///
-/// While we could use `Iterator<Item = Result<Row, ExecutorError>>` directly,
-/// this trait adds query-specific capabilities:
-/// - Access to the output schema (for type checking and column resolution)
-/// - Size hints for query optimization
-/// - Future: Statistics, cost estimates, etc.
-pub trait RowIterator: Iterator<Item = Result<storage::Row, ExecutorError>> {
-    /// Get the schema of rows produced by this iterator
-    ///
-    /// The schema defines the structure and types of columns in output rows.
-    /// It remains constant throughout iteration and must match the schema
-    /// of all rows produced.
-    fn schema(&self) -> &CombinedSchema;
-
-    /// Provide a hint about the number of rows this iterator will produce
-    ///
-    /// This follows the same semantics as `Iterator::size_hint()`:
-    /// - Returns `(lower_bound, upper_bound)`
-    /// - `lower_bound` is always `<= actual count <= upper_bound.unwrap_or(usize::MAX)`
-    /// - None for upper_bound means "unknown" or "unbounded"
-    ///
-    /// These hints can be used for:
-    /// - Allocating appropriately-sized buffers
-    /// - Choosing between nested loop vs hash join
-    /// - Query planning and optimization
-    ///
-    /// The default implementation delegates to the underlying Iterator::size_hint()
-    fn row_size_hint(&self) -> (usize, Option<usize>) {
-        self.size_hint()
-    }
-}
-
-/// Iterator that scans rows from a materialized table
-///
-/// This is the simplest RowIterator - it wraps a Vec<Row> and produces
-/// rows one at a time. While the input is still materialized, this provides
-/// a uniform iterator interface for composition with other operators.
-///
-/// # Example
-///
-/// ```rust,ignore
-/// let rows = vec![row1, row2, row3];
-/// let scan = TableScanIterator::new(schema, rows);
-/// for row in scan {
-///     println!("{:?}", row?);
-/// }
-/// ```
-pub struct TableScanIterator {
-    schema: CombinedSchema,
-    rows: std::vec::IntoIter<storage::Row>,
-    total_count: usize,
-}
-
-impl TableScanIterator {
-    /// Create a new table scan iterator from a schema and materialized rows
-    ///
-    /// # Arguments
-    /// * `schema` - The schema describing the structure of rows
-    /// * `rows` - The rows to iterate over (consumed and moved)
-    pub fn new(schema: CombinedSchema, rows: Vec<storage::Row>) -> Self {
-        let total_count = rows.len();
-        Self { schema, rows: rows.into_iter(), total_count }
-    }
-}
-
-impl Iterator for TableScanIterator {
-    type Item = Result<storage::Row, ExecutorError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.rows.next().map(Ok)
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.rows.size_hint()
-    }
-}
-
-impl RowIterator for TableScanIterator {
-    fn schema(&self) -> &CombinedSchema {
-        &self.schema
-    }
-
-    fn row_size_hint(&self) -> (usize, Option<usize>) {
-        let (lower, upper) = self.size_hint();
-        (lower, upper)
-    }
-}
-
-/// Iterator that filters rows based on a predicate expression
-///
-/// This iterator wraps another RowIterator and only yields rows that
-/// satisfy the given predicate. Rows are evaluated lazily - predicates
-/// are only evaluated for rows that are actually consumed.
-///
-/// # Predicate Evaluation
-///
-/// The predicate must evaluate to a boolean (or truthy value for SQLLogicTest compat):
-/// - `true` or non-zero number: row passes
-/// - `false`, `0`, or `NULL`: row is filtered out
-///
-/// # Example
-///
-/// ```rust,ignore
-/// let scan = TableScanIterator::new(schema.clone(), rows);
-/// let predicate = Expression::Binary {
-///     left: Box::new(Expression::ColumnRef { column: "age".to_string() }),
-///     op: BinaryOp::GreaterThan,
-///     right: Box::new(Expression::Literal(SqlValue::Integer(18))),
-/// };
-/// let filter = FilterIterator::new(scan, predicate, evaluator);
-///
-/// // Only rows where age > 18 are yielded
-/// for row in filter {
-///     println!("{:?}", row?);
-/// }
-/// ```
-pub struct FilterIterator<'a, I: RowIterator> {
-    source: I,
-    predicate: ast::Expression,
-    evaluator: CombinedExpressionEvaluator<'a>,
-}
-
-impl<'a, I: RowIterator> FilterIterator<'a, I> {
-    /// Create a new filter iterator
-    ///
-    /// # Arguments
-    /// * `source` - The source iterator to filter
-    /// * `predicate` - The expression to evaluate for each row
-    /// * `evaluator` - The evaluator to use for predicate evaluation
-    pub fn new(source: I, predicate: ast::Expression, evaluator: CombinedExpressionEvaluator<'a>) -> Self {
-        Self { source, predicate, evaluator }
-    }
-
-    /// Check if a value is "truthy" according to SQL/SQLLogicTest semantics
-    ///
-    /// Returns:
-    /// - `Ok(true)` if the value is truthy (row should be included)
-    /// - `Ok(false)` if the value is falsy (row should be filtered out)
-    /// - `Err(...)` if the value is not a valid predicate type
-    fn is_truthy(value: &types::SqlValue) -> Result<bool, ExecutorError> {
-        match value {
-            types::SqlValue::Boolean(b) => Ok(*b),
-            types::SqlValue::Null => Ok(false), // NULL is falsy
-            // SQLLogicTest compatibility: treat integers as truthy/falsy
-            types::SqlValue::Integer(0)
-            | types::SqlValue::Smallint(0)
-            | types::SqlValue::Bigint(0) => Ok(false),
-            types::SqlValue::Integer(_) | types::SqlValue::Smallint(_) | types::SqlValue::Bigint(_) => {
-                Ok(true)
-            }
-            types::SqlValue::Float(f) | types::SqlValue::Real(f) if *f == 0.0 => Ok(false),
-            types::SqlValue::Float(_) | types::SqlValue::Real(_) => Ok(true),
-            types::SqlValue::Double(f) if *f == 0.0 => Ok(false),
-            types::SqlValue::Double(_) => Ok(true),
-            other => Err(ExecutorError::InvalidWhereClause(format!(
-                "Filter expression must evaluate to boolean, got: {:?}",
-                other
-            ))),
-        }
-    }
-}
-
-// Implement RowIterator for Box<dyn RowIterator> to allow boxing
-// Note: Box<T> already implements Iterator if T implements Iterator,
-// so we only need to implement the RowIterator trait methods
-impl<'a> RowIterator for Box<dyn RowIterator + 'a> {
-    fn schema(&self) -> &CombinedSchema {
-        (**self).schema()
-    }
-
-    fn row_size_hint(&self) -> (usize, Option<usize>) {
-        (**self).row_size_hint()
-    }
-}
-
-impl<'a, I: RowIterator> Iterator for FilterIterator<'a, I> {
-    type Item = Result<storage::Row, ExecutorError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        loop {
-            // Get the next row from source
-            let row = match self.source.next()? {
-                Ok(row) => row,
-                Err(e) => return Some(Err(e)),
-            };
-
-            // Evaluate the predicate for this row
-            match self.evaluator.eval(&self.predicate, &row) {
-                Ok(value) => match Self::is_truthy(&value) {
-                    Ok(true) => return Some(Ok(row)),  // Row passes filter
-                    Ok(false) => continue,              // Row filtered out, try next
-                    Err(e) => return Some(Err(e)),     // Type error
-                },
-                Err(e) => return Some(Err(e)), // Evaluation error
-            }
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        // Filter can only reduce the count, never increase it
-        // Lower bound is 0 (might filter everything)
-        // Upper bound is the source's upper bound (might filter nothing)
-        let (_, upper) = self.source.size_hint();
-        (0, upper)
-    }
-}
-
-impl<'a, I: RowIterator> RowIterator for FilterIterator<'a, I> {
-    fn schema(&self) -> &CombinedSchema {
-        self.source.schema()
-    }
-
-    fn row_size_hint(&self) -> (usize, Option<usize>) {
-        // Same logic as size_hint
-        let (_, upper) = self.source.row_size_hint();
-        (0, upper)
-    }
-}
-
-/// Iterator that projects (transforms) rows using a projection function
-///
-/// This iterator wraps another RowIterator and applies a transformation
-/// to each row. The transformation can change the row's values and/or schema
-/// (e.g., selecting specific columns, computing expressions, renaming columns).
-///
-/// # Example
-///
-/// ```rust,ignore
-/// let scan = TableScanIterator::new(schema, rows);
-///
-/// // Project to select only certain columns
-/// let project_fn = |row: storage::Row| -> Result<storage::Row, ExecutorError> {
-///     Ok(storage::Row::new(vec![row.values[0].clone(), row.values[2].clone()]))
-/// };
-///
-/// let projected_schema = CombinedSchema::new(/* ... */);
-/// let project = ProjectionIterator::new(scan, projected_schema, project_fn);
-///
-/// for row in project {
-///     println!("{:?}", row?);  // Only contains projected columns
-/// }
-/// ```
-pub struct ProjectionIterator<I, F>
-where
-    I: RowIterator,
-    F: FnMut(storage::Row) -> Result<storage::Row, ExecutorError>,
-{
-    source: I,
-    schema: CombinedSchema,
-    projection_fn: F,
-}
-
-impl<I, F> ProjectionIterator<I, F>
-where
-    I: RowIterator,
-    F: FnMut(storage::Row) -> Result<storage::Row, ExecutorError>,
-{
-    /// Create a new projection iterator
-    ///
-    /// # Arguments
-    /// * `source` - The source iterator to project
-    /// * `schema` - The schema of rows after projection
-    /// * `projection_fn` - The function to transform each row
-    pub fn new(source: I, schema: CombinedSchema, projection_fn: F) -> Self {
-        Self { source, schema, projection_fn }
-    }
-}
-
-impl<I, F> Iterator for ProjectionIterator<I, F>
-where
-    I: RowIterator,
-    F: FnMut(storage::Row) -> Result<storage::Row, ExecutorError>,
-{
-    type Item = Result<storage::Row, ExecutorError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let row = match self.source.next()? {
-            Ok(row) => row,
-            Err(e) => return Some(Err(e)),
-        };
-
-        Some((self.projection_fn)(row))
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        // Projection doesn't change row count
-        self.source.size_hint()
-    }
-}
-
-impl<I, F> RowIterator for ProjectionIterator<I, F>
-where
-    I: RowIterator,
-    F: FnMut(storage::Row) -> Result<storage::Row, ExecutorError>,
-{
-    fn schema(&self) -> &CombinedSchema {
-        &self.schema
-    }
-
-    fn row_size_hint(&self) -> (usize, Option<usize>) {
-        // Projection doesn't change row count
-        self.source.row_size_hint()
-    }
-}
-
-/// Iterator that performs lazy nested loop join between two row sources
-///
-/// This iterator streams through the left side while materializing the right side.
-/// For each left row, it iterates through all right rows, checking the join condition
-/// and yielding matching combinations.
-///
-/// # Join Types Supported
-///
-/// - **INNER**: Only matching row pairs
-/// - **CROSS**: All combinations (condition ignored)
-/// - **LEFT OUTER**: All left rows + NULLs for unmatched
-/// - **RIGHT OUTER**: All right rows + NULLs for unmatched (requires full left scan first)
-/// - **FULL OUTER**: Combination of LEFT and RIGHT
-///
-/// # Current Limitation
-///
-/// The right side must be fully materialized (Vec<Row>). This is a pragmatic compromise:
-/// - Nested loop joins inherently need to scan the right side multiple times
-/// - True streaming would require a hash join or other algorithm
-/// - This still provides memory benefits for the left side (which can be very large)
-///
-/// # Example
-///
-/// ```rust,ignore
-/// let left_iter = TableScanIterator::new(left_schema, left_rows);
-/// let right_rows = vec![/* right side rows */];
-///
-/// let join = LazyNestedLoopJoin::new(
-///     left_iter,
-///     right_schema,
-///     right_rows,
-///     ast::JoinType::Inner,
-///     Some(condition),
-///     evaluator,
-/// );
-///
-/// // Only materializes left rows on-demand
-/// for row in join.take(10) {
-///     println!("{:?}", row?);
-/// }
-/// ```
-pub struct LazyNestedLoopJoin<I: RowIterator> {
-    /// Left side iterator (streaming)
-    left: I,
-    /// Right side schema (needed for combined schema)
-    right_schema: CombinedSchema,
-    /// Right side rows (materialized)
-    right_rows: Vec<storage::Row>,
-    /// Type of join
-    join_type: ast::JoinType,
-    /// Optional join condition
-    condition: Option<ast::Expression>,
-    /// Combined schema for output rows
-    combined_schema: CombinedSchema,
-
-    // State for iteration
-    /// Current left row being processed
-    current_left: Option<storage::Row>,
-    /// Current position in right rows
-    right_index: usize,
-    /// For LEFT/FULL OUTER: tracks if current left row has any matches
-    current_left_matched: bool,
-    /// For RIGHT/FULL OUTER: tracks which right rows have been matched
-    right_matched: Vec<bool>,
-    /// For RIGHT/FULL OUTER: after left exhausted, yield unmatched right rows
-    left_exhausted: bool,
-    /// For RIGHT/FULL OUTER: current position in emitting unmatched right rows
-    unmatched_right_index: usize,
-}
-
-impl<I: RowIterator> LazyNestedLoopJoin<I> {
-    /// Create a new lazy nested loop join iterator
-    ///
-    /// # Arguments
-    /// * `left` - Iterator for left side rows (streaming)
-    /// * `right_schema` - Schema for right side
-    /// * `right_rows` - Materialized right side rows
-    /// * `join_type` - Type of join (INNER, LEFT, RIGHT, FULL, CROSS)
-    /// * `condition` - Optional join condition expression
-    pub fn new(
-        left: I,
-        right_schema: CombinedSchema,
-        right_rows: Vec<storage::Row>,
-        join_type: ast::JoinType,
-        condition: Option<ast::Expression>,
-    ) -> Self {
-        // Build combined schema (left columns + right columns)
-        let left_schema = left.schema().clone();
-
-        // Manually combine two CombinedSchema instances
-        let mut table_schemas = left_schema.table_schemas.clone();
-        let left_total = left_schema.total_columns;
-        let mut right_total = 0;
-
-        // Add all right-side tables with adjusted start indices
-        for (table_name, (start_idx, schema)) in right_schema.table_schemas.iter() {
-            let adjusted_start = left_total + start_idx;
-            table_schemas.insert(table_name.clone(), (adjusted_start, schema.clone()));
-            right_total += schema.columns.len();
-        }
-
-        let combined_schema = CombinedSchema {
-            table_schemas,
-            total_columns: left_total + right_total,
-        };
-
-        let right_count = right_rows.len();
-
-        Self {
-            left,
-            right_schema,
-            right_rows,
-            join_type,
-            condition,
-            combined_schema,
-            current_left: None,
-            right_index: 0,
-            current_left_matched: false,
-            right_matched: vec![false; right_count],
-            left_exhausted: false,
-            unmatched_right_index: 0,
-        }
-    }
-
-    /// Combine two rows into a single row (left + right)
-    #[inline]
-    fn combine_rows(left: &storage::Row, right: &storage::Row) -> storage::Row {
-        let mut values = Vec::with_capacity(left.values.len() + right.values.len());
-        values.extend_from_slice(&left.values);
-        values.extend_from_slice(&right.values);
-        storage::Row::new(values)
-    }
-
-    /// Create a row with NULL values for outer join
-    #[inline]
-    fn null_row(column_count: usize) -> Vec<types::SqlValue> {
-        vec![types::SqlValue::Null; column_count]
-    }
-
-    /// Check if join condition is satisfied for a combined row
-    fn check_condition(&self, combined_row: &storage::Row) -> Result<bool, ExecutorError> {
-        match &self.condition {
-            None => Ok(true), // No condition means all rows match (CROSS JOIN)
-            Some(expr) => {
-                // Create evaluator on-the-fly for the combined schema
-                let database = storage::Database::new(); // Empty database for column comparisons
-                let evaluator = CombinedExpressionEvaluator::with_database(&self.combined_schema, &database);
-                let value = evaluator.eval(expr, combined_row)?;
-                // Use same truthy logic as FilterIterator
-                match value {
-                    types::SqlValue::Boolean(b) => Ok(b),
-                    types::SqlValue::Null => Ok(false),
-                    types::SqlValue::Integer(0)
-                    | types::SqlValue::Smallint(0)
-                    | types::SqlValue::Bigint(0) => Ok(false),
-                    types::SqlValue::Integer(_)
-                    | types::SqlValue::Smallint(_)
-                    | types::SqlValue::Bigint(_) => Ok(true),
-                    types::SqlValue::Float(f) | types::SqlValue::Real(f) if f == 0.0 => Ok(false),
-                    types::SqlValue::Float(_) | types::SqlValue::Real(_) => Ok(true),
-                    types::SqlValue::Double(f) if f == 0.0 => Ok(false),
-                    types::SqlValue::Double(_) => Ok(true),
-                    other => Err(ExecutorError::InvalidWhereClause(format!(
-                        "Join condition must evaluate to boolean, got: {:?}",
-                        other
-                    ))),
-                }
-            }
-        }
-    }
-}
-
-impl<I: RowIterator> Iterator for LazyNestedLoopJoin<I> {
-    type Item = Result<storage::Row, ExecutorError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        // Handle RIGHT/FULL OUTER: after left exhausted, emit unmatched right rows
-        if self.left_exhausted {
-            match self.join_type {
-                ast::JoinType::RightOuter | ast::JoinType::FullOuter => {
-                    while self.unmatched_right_index < self.right_rows.len() {
-                        let idx = self.unmatched_right_index;
-                        self.unmatched_right_index += 1;
-
-                        if !self.right_matched[idx] {
-                            // Emit: NULL + right_row
-                            let left_null_count = self.combined_schema.table_schemas.values()
-                                .map(|(_, s)| s.columns.len())
-                                .sum::<usize>() - self.right_schema.table_schemas.values()
-                                .map(|(_, s)| s.columns.len())
-                                .sum::<usize>();
-
-                            let mut values = Self::null_row(left_null_count);
-                            values.extend_from_slice(&self.right_rows[idx].values);
-                            return Some(Ok(storage::Row::new(values)));
-                        }
-                    }
-                    return None; // Done with unmatched right rows
-                }
-                _ => return None, // Other join types done when left exhausted
-            }
-        }
-
-        loop {
-            // If no current left row, fetch the next one
-            if self.current_left.is_none() {
-                match self.left.next() {
-                    Some(Ok(row)) => {
-                        self.current_left = Some(row);
-                        self.right_index = 0;
-                        self.current_left_matched = false;
-                    }
-                    Some(Err(e)) => return Some(Err(e)),
-                    None => {
-                        // Left side exhausted
-                        self.left_exhausted = true;
-
-                        // For RIGHT/FULL OUTER, continue to emit unmatched right rows
-                        match self.join_type {
-                            ast::JoinType::RightOuter | ast::JoinType::FullOuter => {
-                                return self.next(); // Recurse to handle unmatched right rows
-                            }
-                            _ => return None,
-                        }
-                    }
-                }
-            }
-
-            // We have a left row - iterate through right rows
-            let left_row = self.current_left.as_ref().unwrap();
-
-            while self.right_index < self.right_rows.len() {
-                let right_row = &self.right_rows[self.right_index];
-                let right_idx = self.right_index;
-                self.right_index += 1;
-
-                // Combine rows for condition check
-                let combined_row = Self::combine_rows(left_row, right_row);
-
-                // Check join condition
-                match self.check_condition(&combined_row) {
-                    Ok(true) => {
-                        // Match found!
-                        self.current_left_matched = true;
-                        self.right_matched[right_idx] = true;
-                        return Some(Ok(combined_row));
-                    }
-                    Ok(false) => {
-                        // No match, continue to next right row
-                        continue;
-                    }
-                    Err(e) => return Some(Err(e)),
-                }
-            }
-
-            // Finished all right rows for this left row
-            // Handle LEFT/FULL OUTER: emit left + NULLs if no matches
-            if !self.current_left_matched {
-                match self.join_type {
-                    ast::JoinType::LeftOuter | ast::JoinType::FullOuter => {
-                        let right_null_count = self.right_schema.table_schemas.values()
-                            .map(|(_, s)| s.columns.len())
-                            .sum::<usize>();
-
-                        let mut values = left_row.values.clone();
-                        values.extend(Self::null_row(right_null_count));
-
-                        self.current_left = None; // Move to next left row
-                        return Some(Ok(storage::Row::new(values)));
-                    }
-                    _ => {
-                        // INNER/CROSS/RIGHT: no match means don't emit
-                        self.current_left = None; // Move to next left row
-                        continue;
-                    }
-                }
-            }
-
-            // Move to next left row
-            self.current_left = None;
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let (left_lower, left_upper) = self.left.size_hint();
-        let right_count = self.right_rows.len();
-
-        match self.join_type {
-            ast::JoinType::Inner => {
-                // At minimum 0 (no matches), at most left * right
-                (0, left_upper.and_then(|l| l.checked_mul(right_count)))
-            }
-            ast::JoinType::Cross => {
-                // Exactly left * right
-                let lower = left_lower.saturating_mul(right_count);
-                let upper = left_upper.and_then(|l| l.checked_mul(right_count));
-                (lower, upper)
-            }
-            ast::JoinType::LeftOuter | ast::JoinType::FullOuter => {
-                // At least as many as left side
-                (left_lower, None) // Upper bound is hard to compute
-            }
-            ast::JoinType::RightOuter => {
-                // At least as many as right side
-                (right_count, None)
-            }
-        }
-    }
-}
-
-impl<I: RowIterator> RowIterator for LazyNestedLoopJoin<I> {
-    fn schema(&self) -> &CombinedSchema {
-        &self.combined_schema
-    }
-
-    fn row_size_hint(&self) -> (usize, Option<usize>) {
-        self.size_hint()
-    }
-}
-
-// ============================================================================
-// Phase C: Integration Strategy & Status
-// ============================================================================
-//
-// ## PHASE C STATUS: PROOF-OF-CONCEPT COMPLETE ✓
-//
-// What's Complete:
-// - ✅ Core iterator infrastructure (TableScanIterator, FilterIterator, ProjectionIterator)
-// - ✅ Lazy join execution (LazyNestedLoopJoin supporting all SQL join types)
-// - ✅ Evaluator bug fixed (CSE cache was incorrectly caching column references)
-// - ✅ End-to-end pipeline validated (19/19 tests passing)
-// - ✅ Integration strategy documented below
-// - ✅ Materialization decision logic (can_use_iterator_execution in nonagg.rs)
-//
-// ## Integration Strategy for Phase D (Production)
-//
-// ### Step 1: Add Iterator Execution Path (nonagg.rs)
-//
-// ```ignore
-// pub(super) fn execute_without_aggregation(...) -> Result<Vec<Row>, ExecutorError> {
-//     // Decision point: simple queries use iterators
-//     if Self::can_use_iterator_execution(stmt) {
-//         return self.execute_with_iterators(stmt, from_result);
-//     }
-//
-//     // Complex queries use existing materialized path
-//     // (ORDER BY, DISTINCT, window functions)
-//     // ... existing code ...
-// }
-// ```
-//
-// ### Step 2: Implement Iterator Execution (demonstrated by tests below)
-//
-// The test_phase_c_proof_of_concept_*() functions demonstrate the complete pattern:
-// 1. Create TableScanIterator from FROM results
-// 2. Chain FilterIterator for WHERE clause
-// 3. Apply LIMIT via .take(n) for early termination
-// 4. Materialize only final results via .collect()
-// 5. Project columns on materialized rows
-//
-// ### Step 3: Benefits (Validated by Tests)
-//
-// - **Memory**: Only final result set is materialized (not intermediate JOINs)
-// - **Performance**: LIMIT 10 on 1000 rows processes only 10 (not 1000!)
-// - **Composability**: Iterator chain naturally (scan → filter → join → limit)
-//
-// ## Next: Phase D - Production Integration
-//
-// 1. Refactor scan.rs to optionally return Box<dyn RowIterator>
-// 2. Implement execute_with_iterators() using pattern from tests
-// 3. Add benchmarks (iterator vs materialized execution)
-// 4. Expand to more query types (currently simple SELECT/WHERE/LIMIT)
-
-/// Proof-of-concept function demonstrating iterator-based query execution
-///
-/// This function shows how to build an iterator pipeline for simple SELECT queries
-/// (without ORDER BY, DISTINCT, or window functions). It serves as a template for
-/// full integration into the executor.
-///
-/// # Pipeline Construction
-///
-/// The pipeline is built in stages:
-/// 1. **FROM**: Start with TableScanIterator or LazyNestedLoopJoin
-/// 2. **WHERE**: Add FilterIterator for predicates
-/// 3. **SELECT**: Add ProjectionIterator for column selection
-/// 4. **LIMIT**: Use standard `.take(n)` for early termination
-///
-/// # Example Usage
-///
-/// ```rust,ignore
-/// // Build iterator for: SELECT name, age FROM users WHERE age > 18 LIMIT 10
-/// let iterator = build_simple_query_iterator(
-///     users_schema,
-///     users_rows,
-///     Some(age_gt_18_expr),      // WHERE age > 18
-///     projection_fn,              // SELECT name, age
-///     Some(10),                   // LIMIT 10
-/// )?;
-///
-/// // Consume results lazily
-/// let results: Vec<_> = iterator.collect::<Result<Vec<_>, _>>()?;
-/// ```
-///
-/// # Limitations
-///
-/// This proof-of-concept does NOT handle:
-/// - ORDER BY (requires materialization for sorting)
-/// - DISTINCT (requires materialization for deduplication)
-/// - Window functions (requires materialization for partitioning)
-/// - Aggregation (requires materialization for grouping)
-///
-/// For queries with these features, the executor must materialize the iterator
-/// before applying the operation.
-///
-/// # Full Integration Path (Phase C Continuation)
-///
-/// To fully integrate this into the executor:
-///
-/// 1. **Modify `execute_from()` signature**:
-///    ```rust
-///    fn execute_from() -> Result<Box<dyn RowIterator>, ExecutorError>
-///    ```
-///
-/// 2. **Add materialization decision logic**:
-///    ```rust
-///    fn needs_materialization(stmt: &SelectStmt) -> bool {
-///        stmt.order_by.is_some()
-///            || stmt.distinct
-///            || has_window_functions(&stmt.select_list)
-///            || has_aggregates(&stmt.select_list)
-///    }
-///    ```
-///
-/// 3. **Hybrid execution in `execute_without_aggregation()`**:
-///    ```rust
-///    let iter = build_query_iterator(from_result)?;
-///
-///    if needs_materialization(stmt) {
-///        let rows = iter.collect::<Result<Vec<_>, _>>()?;
-///        apply_order_by(rows, &stmt.order_by)
-///    } else {
-///        iter.take(stmt.limit.unwrap_or(usize::MAX))
-///            .collect::<Result<Vec<_>, _>>()
-///    }
-///    ```
-///
-/// 4. **Update all FROM clause execution**:
-///    - `execute_from_clause()` returns iterator
-///    - `nested_loop_join()` uses `LazyNestedLoopJoin`
-///    - `execute_table_scan()` uses `TableScanIterator`
-///
-/// This proof-of-concept validates the approach and provides a clear path forward.
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use storage::Row;
-    use types::SqlValue;
-
-    /// Helper to create a simple schema for testing
-    fn test_schema() -> CombinedSchema {
-        let table_schema = catalog::TableSchema::new(
-            "test".to_string(),
-            vec![catalog::ColumnSchema::new("id".to_string(), types::DataType::Integer, false)],
-        );
-        CombinedSchema::from_table("test".to_string(), table_schema)
-    }
-
-    #[test]
-    fn test_table_scan_iterator_empty() {
-        let schema = test_schema();
-        let rows = vec![];
-        let mut iter = TableScanIterator::new(schema.clone(), rows);
-
-        // Can't use assert_eq! on schema (no PartialEq), just verify it's not null
-        assert!(iter.schema().table_schemas.len() >= 0);
-        assert_eq!(iter.next(), None);
-        assert_eq!(iter.size_hint(), (0, Some(0)));
-    }
-
-    #[test]
-    fn test_table_scan_iterator_with_rows() {
-        let schema = test_schema();
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1)]),
-            Row::new(vec![SqlValue::Integer(2)]),
-            Row::new(vec![SqlValue::Integer(3)]),
-        ];
-        let mut iter = TableScanIterator::new(schema.clone(), rows);
-
-        // Can't use assert_eq! on schema (no PartialEq), just verify it's not null
-        assert!(iter.schema().table_schemas.len() >= 0);
-        assert_eq!(iter.size_hint(), (3, Some(3)));
-
-        assert_eq!(iter.next().unwrap().unwrap().values, vec![SqlValue::Integer(1)]);
-        assert_eq!(iter.next().unwrap().unwrap().values, vec![SqlValue::Integer(2)]);
-        assert_eq!(iter.next().unwrap().unwrap().values, vec![SqlValue::Integer(3)]);
-        assert_eq!(iter.next(), None);
-    }
-
-    #[test]
-    fn test_filter_iterator_all_pass() {
-        let schema = test_schema();
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1)]),
-            Row::new(vec![SqlValue::Integer(2)]),
-            Row::new(vec![SqlValue::Integer(3)]),
-        ];
-        let scan = TableScanIterator::new(schema.clone(), rows);
-
-        // Predicate that always returns true
-        let predicate = ast::Expression::Literal(SqlValue::Boolean(true));
-        let evaluator = CombinedExpressionEvaluator::new(&schema);
-        let mut filter = FilterIterator::new(scan, predicate, evaluator);
-
-        // Can't use assert_eq! on schema (no PartialEq), just verify it's not null
-        assert!(filter.schema().table_schemas.len() >= 0);
-        assert_eq!(filter.next().unwrap().unwrap().values, vec![SqlValue::Integer(1)]);
-        assert_eq!(filter.next().unwrap().unwrap().values, vec![SqlValue::Integer(2)]);
-        assert_eq!(filter.next().unwrap().unwrap().values, vec![SqlValue::Integer(3)]);
-        assert_eq!(filter.next(), None);
-    }
-
-    #[test]
-    fn test_filter_iterator_none_pass() {
-        let schema = test_schema();
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1)]),
-            Row::new(vec![SqlValue::Integer(2)]),
-            Row::new(vec![SqlValue::Integer(3)]),
-        ];
-        let scan = TableScanIterator::new(schema.clone(), rows);
-
-        // Predicate that always returns false
-        let predicate = ast::Expression::Literal(SqlValue::Boolean(false));
-        let evaluator = CombinedExpressionEvaluator::new(&schema);
-        let mut filter = FilterIterator::new(scan, predicate, evaluator);
-
-        // Can't use assert_eq! on schema (no PartialEq), just verify it's not null
-        assert!(filter.schema().table_schemas.len() >= 0);
-        assert_eq!(filter.next(), None);
-    }
-
-    #[test]
-    fn test_filter_iterator_null_is_falsy() {
-        let schema = test_schema();
-        let rows = vec![Row::new(vec![SqlValue::Integer(1)])];
-        let scan = TableScanIterator::new(schema.clone(), rows);
-
-        // Predicate that returns NULL (should filter out)
-        let predicate = ast::Expression::Literal(SqlValue::Null);
-        let evaluator = CombinedExpressionEvaluator::new(&schema);
-        let mut filter = FilterIterator::new(scan, predicate, evaluator);
-
-        assert_eq!(filter.next(), None);
-    }
-
-    #[test]
-    fn test_filter_iterator_integer_truthy() {
-        let schema = test_schema();
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1)]),
-            Row::new(vec![SqlValue::Integer(2)]),
-        ];
-        let scan = TableScanIterator::new(schema.clone(), rows);
-
-        // Predicate that returns non-zero integer (truthy)
-        let predicate = ast::Expression::Literal(SqlValue::Integer(42));
-        let evaluator = CombinedExpressionEvaluator::new(&schema);
-        let mut filter = FilterIterator::new(scan, predicate, evaluator);
-
-        assert_eq!(filter.next().unwrap().unwrap().values, vec![SqlValue::Integer(1)]);
-        assert_eq!(filter.next().unwrap().unwrap().values, vec![SqlValue::Integer(2)]);
-        assert_eq!(filter.next(), None);
-    }
-
-    #[test]
-    fn test_filter_iterator_zero_is_falsy() {
-        let schema = test_schema();
-        let rows = vec![Row::new(vec![SqlValue::Integer(1)])];
-        let scan = TableScanIterator::new(schema.clone(), rows);
-
-        // Predicate that returns 0 (falsy)
-        let predicate = ast::Expression::Literal(SqlValue::Integer(0));
-        let evaluator = CombinedExpressionEvaluator::new(&schema);
-        let mut filter = FilterIterator::new(scan, predicate, evaluator);
-
-        assert_eq!(filter.next(), None);
-    }
-
-    #[test]
-    fn test_evaluator_direct() {
-        // Direct test of evaluator with column reference
-        let table_schema = catalog::TableSchema::new(
-            "test".to_string(),
-            vec![
-                catalog::ColumnSchema::new("id".to_string(), types::DataType::Integer, false),
-                catalog::ColumnSchema::new("age".to_string(), types::DataType::Integer, false),
-            ],
-        );
-        let schema = CombinedSchema::from_table("test".to_string(), table_schema);
-        let database = storage::Database::new();
-        let evaluator = CombinedExpressionEvaluator::with_database(&schema, &database);
-
-        let predicate = ast::Expression::BinaryOp {
-            left: Box::new(ast::Expression::ColumnRef {
-                table: None,
-                column: "age".to_string(),
-            }),
-            op: ast::BinaryOperator::GreaterThan,
-            right: Box::new(ast::Expression::Literal(SqlValue::Integer(18))),
-        };
-
-        // Test row 1: age=25, should be true (25 > 18)
-        let row1 = Row::new(vec![SqlValue::Integer(1), SqlValue::Integer(25)]);
-        let result1 = evaluator.eval(&predicate, &row1).unwrap();
-        println!("Row 1 (age=25): {:?}", result1);
-
-        // Test row 2: age=17, should be false (17 > 18 is false)
-        let row2 = Row::new(vec![SqlValue::Integer(2), SqlValue::Integer(17)]);
-        let result2 = evaluator.eval(&predicate, &row2).unwrap();
-        println!("Row 2 (age=17): {:?}", result2);
-
-        // Test row 3: age=30, should be true (30 > 18)
-        let row3 = Row::new(vec![SqlValue::Integer(3), SqlValue::Integer(30)]);
-        let result3 = evaluator.eval(&predicate, &row3).unwrap();
-        println!("Row 3 (age=30): {:?}", result3);
-
-        // Verify expected values
-        assert!(matches!(result1, SqlValue::Boolean(true)), "Row 1 should be true, got {:?}", result1);
-        assert!(matches!(result2, SqlValue::Boolean(false)), "Row 2 should be false, got {:?}", result2);
-        assert!(matches!(result3, SqlValue::Boolean(true)), "Row 3 should be true, got {:?}", result3);
-    }
-
-    #[test]
-    fn test_filter_with_column_ref() {
-        // Test filtering with column reference comparison
-        let table_schema = catalog::TableSchema::new(
-            "test".to_string(),
-            vec![
-                catalog::ColumnSchema::new("id".to_string(), types::DataType::Integer, false),
-                catalog::ColumnSchema::new("age".to_string(), types::DataType::Integer, false),
-            ],
-        );
-        let schema = CombinedSchema::from_table("test".to_string(), table_schema);
-
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Integer(25)]),
-            Row::new(vec![SqlValue::Integer(2), SqlValue::Integer(17)]),
-            Row::new(vec![SqlValue::Integer(3), SqlValue::Integer(30)]),
-        ];
-        let scan = TableScanIterator::new(schema.clone(), rows);
-
-        // Predicate: age > 18 (using unqualified column reference)
-        let predicate = ast::Expression::BinaryOp {
-            left: Box::new(ast::Expression::ColumnRef {
-                table: None,  // Try without table qualifier
-                column: "age".to_string(),
-            }),
-            op: ast::BinaryOperator::GreaterThan,
-            right: Box::new(ast::Expression::Literal(SqlValue::Integer(18))),
-        };
-
-        let database = storage::Database::new();
-        let evaluator = CombinedExpressionEvaluator::with_database(&schema, &database);
-        let filter = FilterIterator::new(scan, predicate, evaluator);
-
-        // Collect ALL results to see what's happening
-        let results: Vec<_> = filter.collect::<Result<Vec<_>, _>>().unwrap();
-
-        // Should get 2 results: id 1 (age 25) and id 3 (age 30)
-        // Row 2 (age 17) should be filtered out
-        assert_eq!(results.len(), 2, "Expected 2 results, got {} results: {:?}", results.len(), results);
-
-        assert_eq!(results[0].values, vec![SqlValue::Integer(1), SqlValue::Integer(25)]);
-        assert_eq!(results[1].values, vec![SqlValue::Integer(3), SqlValue::Integer(30)]);
-    }
-
-    #[test]
-    fn test_projection_iterator_identity() {
-        let schema = test_schema();
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1)]),
-            Row::new(vec![SqlValue::Integer(2)]),
-        ];
-        let scan = TableScanIterator::new(schema.clone(), rows);
-
-        // Identity projection (no-op)
-        let project_fn = |row: Row| Ok(row);
-        let mut project = ProjectionIterator::new(scan, schema.clone(), project_fn);
-
-        // Can't use assert_eq! on schema (no PartialEq), just verify it's not null
-        assert!(project.schema().table_schemas.len() >= 0);
-        assert_eq!(project.next().unwrap().unwrap().values, vec![SqlValue::Integer(1)]);
-        assert_eq!(project.next().unwrap().unwrap().values, vec![SqlValue::Integer(2)]);
-        assert_eq!(project.next(), None);
-    }
-
-    #[test]
-    fn test_projection_iterator_transform() {
-        let schema = test_schema();
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1)]),
-            Row::new(vec![SqlValue::Integer(2)]),
-        ];
-        let scan = TableScanIterator::new(schema.clone(), rows);
-
-        // Double each value
-        let project_fn = |mut row: Row| {
-            if let SqlValue::Integer(n) = row.values[0] {
-                row.values[0] = SqlValue::Integer(n * 2);
-            }
-            Ok(row)
-        };
-        let mut project = ProjectionIterator::new(scan, schema.clone(), project_fn);
-
-        assert_eq!(project.next().unwrap().unwrap().values, vec![SqlValue::Integer(2)]);
-        assert_eq!(project.next().unwrap().unwrap().values, vec![SqlValue::Integer(4)]);
-        assert_eq!(project.next(), None);
-    }
-
-    #[test]
-    fn test_chained_iterators() {
-        let schema = test_schema();
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1)]),
-            Row::new(vec![SqlValue::Integer(2)]),
-            Row::new(vec![SqlValue::Integer(3)]),
-            Row::new(vec![SqlValue::Integer(4)]),
-        ];
-        let scan = TableScanIterator::new(schema.clone(), rows);
-
-        // Filter for values > 1
-        let predicate = ast::Expression::Literal(SqlValue::Integer(1)); // Truthy, for demo
-        let evaluator = CombinedExpressionEvaluator::new(&schema);
-        let filter = FilterIterator::new(scan, predicate, evaluator);
-
-        // Then double each value
-        let project_fn = |mut row: Row| {
-            if let SqlValue::Integer(n) = row.values[0] {
-                row.values[0] = SqlValue::Integer(n * 2);
-            }
-            Ok(row)
-        };
-        let mut project = ProjectionIterator::new(filter, schema.clone(), project_fn);
-
-        // Should get all rows (filter passes all), doubled
-        assert_eq!(project.next().unwrap().unwrap().values, vec![SqlValue::Integer(2)]);
-        assert_eq!(project.next().unwrap().unwrap().values, vec![SqlValue::Integer(4)]);
-        assert_eq!(project.next().unwrap().unwrap().values, vec![SqlValue::Integer(6)]);
-        assert_eq!(project.next().unwrap().unwrap().values, vec![SqlValue::Integer(8)]);
-        assert_eq!(project.next(), None);
-    }
-
-    #[test]
-    fn test_iterator_take_limit() {
-        let schema = test_schema();
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1)]),
-            Row::new(vec![SqlValue::Integer(2)]),
-            Row::new(vec![SqlValue::Integer(3)]),
-            Row::new(vec![SqlValue::Integer(4)]),
-            Row::new(vec![SqlValue::Integer(5)]),
-        ];
-        let scan = TableScanIterator::new(schema.clone(), rows);
-
-        // Use standard iterator take() to implement LIMIT
-        let results: Vec<_> = scan.take(2).collect();
-
-        assert_eq!(results.len(), 2);
-        assert_eq!(results[0].as_ref().unwrap().values, vec![SqlValue::Integer(1)]);
-        assert_eq!(results[1].as_ref().unwrap().values, vec![SqlValue::Integer(2)]);
-    }
-
-    // Helper to create two-table schema for join tests
-    fn test_join_schemas() -> (CombinedSchema, CombinedSchema) {
-        let left_schema = catalog::TableSchema::new(
-            "t1".to_string(),
-            vec![
-                catalog::ColumnSchema::new("id".to_string(), types::DataType::Integer, false),
-                catalog::ColumnSchema::new("value".to_string(), types::DataType::Integer, false),
-            ],
-        );
-        let right_schema = catalog::TableSchema::new(
-            "t2".to_string(),
-            vec![
-                catalog::ColumnSchema::new("id".to_string(), types::DataType::Integer, false),
-                catalog::ColumnSchema::new("data".to_string(), types::DataType::Integer, false),
-            ],
-        );
-        (
-            CombinedSchema::from_table("t1".to_string(), left_schema),
-            CombinedSchema::from_table("t2".to_string(), right_schema),
-        )
-    }
-
-    #[test]
-    fn test_lazy_nested_loop_join_cross() {
-        let (left_schema, right_schema) = test_join_schemas();
-
-        let left_rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Integer(10)]),
-            Row::new(vec![SqlValue::Integer(2), SqlValue::Integer(20)]),
-        ];
-        let right_rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Integer(100)]),
-            Row::new(vec![SqlValue::Integer(2), SqlValue::Integer(200)]),
-        ];
-
-        let left_iter = TableScanIterator::new(left_schema.clone(), left_rows);
-
-        let mut join = LazyNestedLoopJoin::new(
-            left_iter,
-            right_schema,
-            right_rows,
-            ast::JoinType::Cross,
-            None, // No condition for CROSS JOIN
-        );
-
-        // CROSS JOIN: 2 left x 2 right = 4 rows
-        let results: Vec<_> = join.by_ref().collect::<Result<Vec<_>, _>>().unwrap();
-        assert_eq!(results.len(), 4);
-
-        // Check first row: (1, 10, 1, 100)
-        assert_eq!(
-            results[0].values,
-            vec![
-                SqlValue::Integer(1),
-                SqlValue::Integer(10),
-                SqlValue::Integer(1),
-                SqlValue::Integer(100)
-            ]
-        );
-
-        // Check last row: (2, 20, 2, 200)
-        assert_eq!(
-            results[3].values,
-            vec![
-                SqlValue::Integer(2),
-                SqlValue::Integer(20),
-                SqlValue::Integer(2),
-                SqlValue::Integer(200)
-            ]
-        );
-    }
-
-    #[test]
-    fn test_lazy_nested_loop_join_inner_with_condition() {
-        let (left_schema, right_schema) = test_join_schemas();
-
-        let left_rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Integer(10)]),
-            Row::new(vec![SqlValue::Integer(2), SqlValue::Integer(20)]),
-            Row::new(vec![SqlValue::Integer(3), SqlValue::Integer(30)]),
-        ];
-        let right_rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Integer(100)]),
-            Row::new(vec![SqlValue::Integer(2), SqlValue::Integer(200)]),
-        ];
-
-        let left_iter = TableScanIterator::new(left_schema.clone(), left_rows);
-
-        // Condition: t1.id = t2.id (column 0 = column 2)
-        let condition = ast::Expression::BinaryOp {
-            left: Box::new(ast::Expression::ColumnRef {
-                table: Some("t1".to_string()),
-                column: "id".to_string(),
-            }),
-            op: ast::BinaryOperator::Equal,
-            right: Box::new(ast::Expression::ColumnRef {
-                table: Some("t2".to_string()),
-                column: "id".to_string(),
-            }),
-        };
-
-        let mut join = LazyNestedLoopJoin::new(
-            left_iter,
-            right_schema,
-            right_rows,
-            ast::JoinType::Inner,
-            Some(condition),
-        );
-
-        // INNER JOIN with condition: only (1,1) and (2,2) match
-        let results: Vec<_> = join.by_ref().collect::<Result<Vec<_>, _>>().unwrap();
-        assert_eq!(results.len(), 2);
-
-        // (1, 10, 1, 100)
-        assert_eq!(
-            results[0].values,
-            vec![
-                SqlValue::Integer(1),
-                SqlValue::Integer(10),
-                SqlValue::Integer(1),
-                SqlValue::Integer(100)
-            ]
-        );
-
-        // (2, 20, 2, 200)
-        assert_eq!(
-            results[1].values,
-            vec![
-                SqlValue::Integer(2),
-                SqlValue::Integer(20),
-                SqlValue::Integer(2),
-                SqlValue::Integer(200)
-            ]
-        );
-    }
-
-    #[test]
-    fn test_lazy_nested_loop_join_left_outer() {
-        let (left_schema, right_schema) = test_join_schemas();
-
-        let left_rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Integer(10)]),
-            Row::new(vec![SqlValue::Integer(3), SqlValue::Integer(30)]), // No match in right
-        ];
-        let right_rows = vec![Row::new(vec![SqlValue::Integer(1), SqlValue::Integer(100)])];
-
-        let left_iter = TableScanIterator::new(left_schema.clone(), left_rows);
-
-        // Condition: t1.id = t2.id
-        let condition = ast::Expression::BinaryOp {
-            left: Box::new(ast::Expression::ColumnRef {
-                table: Some("t1".to_string()),
-                column: "id".to_string(),
-            }),
-            op: ast::BinaryOperator::Equal,
-            right: Box::new(ast::Expression::ColumnRef {
-                table: Some("t2".to_string()),
-                column: "id".to_string(),
-            }),
-        };
-
-        let mut join = LazyNestedLoopJoin::new(
-            left_iter,
-            right_schema,
-            right_rows,
-            ast::JoinType::LeftOuter,
-            Some(condition),
-        );
-
-        let results: Vec<_> = join.by_ref().collect::<Result<Vec<_>, _>>().unwrap();
-        assert_eq!(results.len(), 2);
-
-        // First row: (1, 10, 1, 100) - match
-        assert_eq!(
-            results[0].values,
-            vec![
-                SqlValue::Integer(1),
-                SqlValue::Integer(10),
-                SqlValue::Integer(1),
-                SqlValue::Integer(100)
-            ]
-        );
-
-        // Second row: (3, 30, NULL, NULL) - no match, left with NULLs
-        assert_eq!(
-            results[1].values,
-            vec![SqlValue::Integer(3), SqlValue::Integer(30), SqlValue::Null, SqlValue::Null]
-        );
-    }
-
-    #[test]
-    fn test_lazy_nested_loop_join_early_termination() {
-        let (left_schema, right_schema) = test_join_schemas();
-
-        // Large left side
-        let left_rows: Vec<_> = (1..=1000)
-            .map(|i| Row::new(vec![SqlValue::Integer(i), SqlValue::Integer(i * 10)]))
-            .collect();
-
-        // Small right side
-        let right_rows: Vec<_> = (1..=10)
-            .map(|i| Row::new(vec![SqlValue::Integer(i), SqlValue::Integer(i * 100)]))
-            .collect();
-
-        let left_iter = TableScanIterator::new(left_schema.clone(), left_rows);
-
-        let join = LazyNestedLoopJoin::new(
-            left_iter,
-            right_schema,
-            right_rows,
-            ast::JoinType::Cross,
-            None,
-        );
-
-        // Only take first 5 rows - should not materialize all 1000 left rows
-        let results: Vec<_> = join.take(5).collect::<Result<Vec<_>, _>>().unwrap();
-        assert_eq!(results.len(), 5);
-
-        // Verify we got the expected rows (first left row with first 5 right rows)
-        for i in 0..5 {
-            assert_eq!(results[i].values[0], SqlValue::Integer(1)); // Left id stays 1
-            assert_eq!(results[i].values[1], SqlValue::Integer(10)); // Left value stays 10
-        }
-    }
-
-    /// Phase C Proof-of-Concept: End-to-End Iterator Pipeline
-    ///
-    /// This test demonstrates how iterators would be used for a complete query:
-    /// SELECT * FROM users WHERE age > 18 LIMIT 10
-    ///
-    /// This validates that:
-    /// 1. TableScanIterator provides the data source
-    /// 2. FilterIterator applies WHERE conditions
-    /// 3. LIMIT works via .take()
-    /// 4. Everything composes naturally
-    #[test]
-    fn test_phase_c_proof_of_concept_full_pipeline() {
-        // Simulated table: users(id, name, age)
-        let schema = catalog::TableSchema::new(
-            "users".to_string(),
-            vec![
-                catalog::ColumnSchema::new("id".to_string(), types::DataType::Integer, false),
-                catalog::ColumnSchema::new("name".to_string(), types::DataType::Varchar { max_length: None }, false),
-                catalog::ColumnSchema::new("age".to_string(), types::DataType::Integer, false),
-            ],
-        );
-        let from_schema = CombinedSchema::from_table("users".to_string(), schema);
-
-        // Test data: 5 users with varying ages
-        let from_rows = vec![
-            Row::new(vec![
-                SqlValue::Integer(1),
-                SqlValue::Varchar("Alice".to_string()),
-                SqlValue::Integer(25),
-            ]),
-            Row::new(vec![
-                SqlValue::Integer(2),
-                SqlValue::Varchar("Bob".to_string()),
-                SqlValue::Integer(17),
-            ]),
-            Row::new(vec![
-                SqlValue::Integer(3),
-                SqlValue::Varchar("Charlie".to_string()),
-                SqlValue::Integer(30),
-            ]),
-            Row::new(vec![
-                SqlValue::Integer(4),
-                SqlValue::Varchar("Diana".to_string()),
-                SqlValue::Integer(16),
-            ]),
-            Row::new(vec![
-                SqlValue::Integer(5),
-                SqlValue::Varchar("Eve".to_string()),
-                SqlValue::Integer(22),
-            ]),
-        ];
-
-        // Stage 1: FROM - Create table scan iterator
-        let scan = TableScanIterator::new(from_schema.clone(), from_rows);
-
-        // Stage 2: WHERE age > 18
-        let where_expr = ast::Expression::BinaryOp {
-            left: Box::new(ast::Expression::ColumnRef {
-                table: Some("users".to_string()),
-                column: "age".to_string(),
-            }),
-            op: ast::BinaryOperator::GreaterThan,
-            right: Box::new(ast::Expression::Literal(SqlValue::Integer(18))),
-        };
-
-        let evaluator = CombinedExpressionEvaluator::new(&from_schema);
-        let filter = FilterIterator::new(scan, where_expr, evaluator);
-
-        // Stage 3: LIMIT 10
-        let limited = filter.take(10);
-
-        // Stage 4: Execute (collect results)
-        let results: Vec<_> = limited.collect::<Result<Vec<_>, _>>().unwrap();
-
-        // Verify results: Should get Alice (25), Charlie (30), Eve (22)
-        // Bob (17) and Diana (16) are filtered out
-        assert_eq!(results.len(), 3);
-
-        assert_eq!(results[0].values[0], SqlValue::Integer(1)); // Alice
-        assert_eq!(results[0].values[2], SqlValue::Integer(25));
-
-        assert_eq!(results[1].values[0], SqlValue::Integer(3)); // Charlie
-        assert_eq!(results[1].values[2], SqlValue::Integer(30));
-
-        assert_eq!(results[2].values[0], SqlValue::Integer(5)); // Eve
-        assert_eq!(results[2].values[2], SqlValue::Integer(22));
-    }
-
-    /// Phase C Proof-of-Concept: Iterator Pipeline with JOIN
-    ///
-    /// This test demonstrates iterator-based execution for a query with JOIN:
-    /// SELECT * FROM orders JOIN customers ON orders.customer_id = customers.id
-    /// WHERE orders.amount > 100 LIMIT 5
-    ///
-    /// This validates that:
-    /// 1. LazyNestedLoopJoin streams through left side (orders)
-    /// 2. JOIN condition is evaluated correctly
-    /// 3. Early termination works (only processes enough to get 5 results)
-    #[test]
-    fn test_phase_c_proof_of_concept_join_pipeline() {
-        // Setup schemas
-        let orders_schema = catalog::TableSchema::new(
-            "orders".to_string(),
-            vec![
-                catalog::ColumnSchema::new("id".to_string(), types::DataType::Integer, false),
-                catalog::ColumnSchema::new("customer_id".to_string(), types::DataType::Integer, false),
-                catalog::ColumnSchema::new("amount".to_string(), types::DataType::Integer, false),
-            ],
-        );
-        let customers_schema = catalog::TableSchema::new(
-            "customers".to_string(),
-            vec![
-                catalog::ColumnSchema::new("id".to_string(), types::DataType::Integer, false),
-                catalog::ColumnSchema::new("name".to_string(), types::DataType::Varchar { max_length: None }, false),
-            ],
-        );
-
-        let orders_combined = CombinedSchema::from_table("orders".to_string(), orders_schema);
-        let customers_combined = CombinedSchema::from_table("customers".to_string(), customers_schema);
-
-        // Test data: 10 orders
-        let orders_rows: Vec<_> = (1..=10)
-            .map(|i| {
-                Row::new(vec![
-                    SqlValue::Integer(i),
-                    SqlValue::Integer((i % 3) + 1), // customer_id cycles 1, 2, 3
-                    SqlValue::Integer(i * 50),       // amount: 50, 100, 150, ...
-                ])
-            })
-            .collect();
-
-        // 3 customers
-        let customers_rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Varchar("Alice".to_string())]),
-            Row::new(vec![SqlValue::Integer(2), SqlValue::Varchar("Bob".to_string())]),
-            Row::new(vec![SqlValue::Integer(3), SqlValue::Varchar("Charlie".to_string())]),
-        ];
-
-        // Stage 1: FROM orders (scan)
-        let orders_scan = TableScanIterator::new(orders_combined.clone(), orders_rows);
-
-        // Stage 2: JOIN customers ON orders.customer_id = customers.id
-        let join_condition = ast::Expression::BinaryOp {
-            left: Box::new(ast::Expression::ColumnRef {
-                table: Some("orders".to_string()),
-                column: "customer_id".to_string(),
-            }),
-            op: ast::BinaryOperator::Equal,
-            right: Box::new(ast::Expression::ColumnRef {
-                table: Some("customers".to_string()),
-                column: "id".to_string(),
-            }),
-        };
-
-        let join = LazyNestedLoopJoin::new(
-            orders_scan,
-            customers_combined.clone(),
-            customers_rows,
-            ast::JoinType::Inner,
-            Some(join_condition),
-        );
-
-        // Stage 3: WHERE orders.amount > 100
-        // Build combined schema for WHERE evaluation
-        let mut combined_tables = orders_combined.table_schemas.clone();
-        for (name, (start_idx, schema)) in customers_combined.table_schemas.iter() {
-            combined_tables.insert(
-                name.clone(),
-                (orders_combined.total_columns + start_idx, schema.clone()),
-            );
-        }
-        let combined_schema = CombinedSchema {
-            table_schemas: combined_tables,
-            total_columns: orders_combined.total_columns + customers_combined.total_columns,
-        };
-
-        let where_expr = ast::Expression::BinaryOp {
-            left: Box::new(ast::Expression::ColumnRef {
-                table: Some("orders".to_string()),
-                column: "amount".to_string(),
-            }),
-            op: ast::BinaryOperator::GreaterThan,
-            right: Box::new(ast::Expression::Literal(SqlValue::Integer(100))),
-        };
-
-        let evaluator = CombinedExpressionEvaluator::new(&combined_schema);
-        let filter = FilterIterator::new(join, where_expr, evaluator);
-
-        // Stage 4: LIMIT 5
-        let limited = filter.take(5);
-
-        // Stage 5: Execute
-        let results: Vec<_> = limited.collect::<Result<Vec<_>, _>>().unwrap();
-
-        // Verify: Should get orders with amount > 100 (orders 3, 4, 5, 6, 7...)
-        assert_eq!(results.len(), 5);
-
-        // Each result should have 5 columns: orders(id, customer_id, amount) + customers(id, name)
-        assert_eq!(results[0].values.len(), 5);
-
-        // First result should be order 3 (amount 150) joined with customer
-        assert_eq!(results[0].values[0], SqlValue::Integer(3)); // order.id
-        assert_eq!(results[0].values[2], SqlValue::Integer(150)); // order.amount
-    }
-}
